src/main/java/com/tdd/katas/dlwob/microservices/vehicleservice/service/PartDataServiceProxy.java
package com.tdd.katas.dlwob.microservices.vehicleservice.service;

import com.tdd.katas.dlwob.microservices.vehicleservice.model.PartData;
import org.springframework.boot.web.client.RestTemplateBuilder;
import org.springframework.stereotype.Component;
import org.springframework.web.client.RestTemplate;

import java.util.List;

@Component
class PartDataServiceProxy {

    private RestTemplate restTemplate;

    public PartDataServiceProxy(RestTemplateBuilder restTemplateBuilder) {
        restTemplate = restTemplateBuilder.build();
    }

    public List<PartData> getPartDataList(String vinCode) {
        return null;
    }

}
